chore(runtime): keep observation files within lint limits
