+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/dispatch_worker_sessions_cutover.go
@@ -19,32 +19,6 @@ import (
 	workerexecution "github.com/portpowered/infinite-you/pkg/services/workers"
 )
 
-func recordDispatchWorkerSessionAssociation(
-	ledger recordings.RuntimeLedger,
-	tick int,
-	dispatchID string,
-	workerSessionID string,
-	requestID string,
-	facts recordings.DispatchWorkerSessionExecutionFacts,
-	eventTime time.Time,
-) {
-	if ledger == nil {
-		return
-	}
-	if recorder, ok := ledger.(recordings.DispatchWorkerSessionAssociationRecorder); ok {
-		recorder.RecordDispatchWorkerSessionAssociationWithExecution(
-			tick,
-			dispatchID,
-			workerSessionID,
-			requestID,
-			facts,
-			eventTime,
-		)
-		return
-	}
-	ledger.RecordDispatchWorkerSessionAssociation(tick, dispatchID, workerSessionID, requestID, eventTime)
-}
-
 // startThroughWorkerSessions is the W4 Runtime dispatch cutover seam. For
 // every resolved dispatch it reserves one stable, control-addressable Worker
 // Session identity, commits
